chore(suite): move CoinControl components to views
